-prepare resip/stack for merge to main  - windows build fixes  - added missing license blocks  - svn eol-style: native  - removed duplicate ConsumerFifoBuffer.hxx (also in rutil)  - added missing files to VS2010 projects  - all projects in main solution now build and link under windows -fixed bug in TuSelector with uninitialized mCongestionManager member
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/branches/b-TKLC-perf_work@9336 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ resip/stack/TransactionUser.hxx
@@ -139,7 +139,7 @@ class TransactionUser
       
       virtual UInt16 getExpectedWait() const
       {
-         return mFifo.expectedWaitTimeMilliSec();
+         return (UInt16)mFifo.expectedWaitTimeMilliSec();
       }
       
       // .bwc. This specifies whether the TU can cope with dropped responses
